test: cover root decorator dep resolution with builtin decorator, drop extra fixture plugin
