--- packages/varlock/src/env-graph/test/plugins/test-plugin-init-args/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@varlock/test-plugin-init-args",
-  "private": true,
-  "description": "plugin used in automated tests - init decorator with args that reference config items",
-  "version": "1.0.0",
-  "type": "module",
-  "exports": {
-    "./plugin": "./plugin.js"
-  }
-}
