+++ packages/varlock/src/env-graph/test/root-decorator-deps.test.ts
@@ -3,31 +3,52 @@ import outdent from 'outdent';
 import { envFilesTest } from './helpers/generic-test';
 
 /**
- * Root decorators (@initAws, @initOnePassword, etc) resolve the items referenced in their
- * args before the normal top-to-bottom resolution pass. Those items may themselves depend
- * on other items, so the whole transitive dependency closure has to be resolved first -
- * otherwise the referenced item never resolves and the failure surfaces as a confusing
- * `Referenced item "X" is not valid` on an unrelated arg. See issue #940.
+ * Root decorators resolve the items referenced in their args before the normal
+ * top-to-bottom resolution pass. Those items may themselves depend on other items, so the
+ * whole transitive dependency closure has to be resolved first - otherwise the referenced
+ * item never resolves and the failure surfaces as a confusing `Referenced item "X" is not
+ * valid` on an unrelated arg. See issue #940.
  */
 describe('root decorator arg dependencies', () => {
-  test('resolves direct deps of init decorator args', envFilesTest({
+  test('resolves direct deps of a root decorator value', envFilesTest({
     envFile: outdent`
-      # @plugin(./plugins/test-plugin-init-args/)
-      # @initTestArgs(value=$PROFILE)
+      # @redactLogs=$SHOULD_REDACT
       # ---
-      PROFILE=static-profile
-      INIT_VALUE=initArgValue()
+      SHOULD_REDACT=false
+    `,
+    expectValues: { SHOULD_REDACT: false },
+  }));
+
+  test('resolves transitive deps of a root decorator value', envFilesTest({
+    envFile: outdent`
+      # @redactLogs=$SHOULD_REDACT
+      # ---
+      SHOULD_REDACT=ifs(eq($DEPLOY_ENV, "dev"), "false", "true")
+      DEPLOY_ENV=dev
     `,
     expectValues: {
-      PROFILE: 'static-profile',
-      INIT_VALUE: 'static-profile',
+      SHOULD_REDACT: 'false',
+      DEPLOY_ENV: 'dev',
     },
   }));
 
-  test('resolves transitive deps of init decorator args', envFilesTest({
+  test('resolves multi-level transitive deps of a root decorator value', envFilesTest({
     envFile: outdent`
-      # @plugin(./plugins/test-plugin-init-args/)
-      # @initTestArgs(value=$PROFILE)
+      # @redactLogs=$SHOULD_REDACT
+      # ---
+      SHOULD_REDACT=concat($REDACT_PREFIX, "alse")
+      REDACT_PREFIX=ifs(eq($DEPLOY_ENV, "dev"), "f", "t")
+      DEPLOY_ENV=dev
+    `,
+    expectValues: { SHOULD_REDACT: 'false' },
+  }));
+
+  // the reported case - a plugin init decorator (ex: @initAws(profile=$AWS_PROFILE)) whose
+  // args reference items that have deps of their own
+  test('plugin init decorator receives the resolved value of a dep with its own deps', envFilesTest({
+    envFile: outdent`
+      # @plugin(./plugins/test-plugin-with-init/)
+      # @initTestPlugin(value=$PROFILE)
       # ---
       PROFILE=ifs(eq($DEPLOY_ENV, "dev"), "profile-dev", "profile-prod")
       DEPLOY_ENV=dev
@@ -35,24 +56,7 @@ describe('root decorator arg dependencies', () => {
     `,
     expectValues: {
       PROFILE: 'profile-dev',
-      DEPLOY_ENV: 'dev',
       INIT_VALUE: 'profile-dev',
-    },
-  }));
-
-  test('resolves multi-level transitive deps of init decorator args', envFilesTest({
-    envFile: outdent`
-      # @plugin(./plugins/test-plugin-init-args/)
-      # @initTestArgs(value=$PROFILE)
-      # ---
-      PROFILE=concat("profile-", $ENV_NAME)
-      ENV_NAME=ifs(eq($DEPLOY_ENV, "dev"), "development", "production")
-      DEPLOY_ENV=dev
-      INIT_VALUE=initArgValue()
-    `,
-    expectValues: {
-      PROFILE: 'profile-development',
-      INIT_VALUE: 'profile-development',
     },
   }));
 });
